:wrench: restructured ttml models
