+++ app/src/main/java/com/shub39/rush/domain/dataclasses/ParsedWord.kt
@@ -1,0 +1,3 @@
+package com.shub39.rush.domain.dataclasses
+
+data class ParsedWord(val text: String, val startTime: Double, val endTime: Double)
