+++ app/src/main/java/com/shub39/rush/domain/dataclasses/SpanInfo.kt
@@ -1,0 +1,8 @@
+package com.shub39.rush.domain.dataclasses
+
+data class SpanInfo(
+    val text: String,
+    val startTime: Double,
+    val endTime: Double,
+    val hasTrailingSpace: Boolean,
+)
